requirements.txt
statsmodels==0.11.1
pandas==1.1.5
numpy==1.19.5
matplotlib==3.3.4
seaborn==0.10.0
joblib==1.0.1
PyYAML==5.3.1
scikit_learn==0.24.1
